+++ packages/react/src/components/header/index.ts
@@ -8,6 +8,8 @@ export {
   HeaderTitle,
   headerNavLinkVariants,
   headerVariants,
+  useHeaderMenu,
+  type HeaderMenu,
   type HeaderMenuButtonProps,
   type HeaderNavLinkProps,
   type HeaderNavProps,